// initial-source/funky-export/home-ubuntu/disco/front-end/src/types/prizeItem.ts
export type prizeItem =  {
    ranking: string;
    probability: string;
    real_probability: string;
    fake_probability: string;
    tokenName: string;
    price: string;
    symbol: string;
    earned_pts: string;
    ca: string;
    telegram: string;
    twitter: string;
    listed_dex: string;
    defaultImage: string;
}